# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SOWISO Etiquette API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scenarios:
    get:
      operationId: getScenarios
      tags: [scenarios]
      summary: Get training scenarios filtered by region, pillar, and difficulty
      parameters:
        - name: region_code
          in: query
          required: false
          schema:
            type: string
        - name: pillar
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: difficulty_level
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: difficulty_max
          in: query
          required: false
          description: Return scenarios with difficulty_level up to this value (inclusive)
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: age_group
          in: query
          required: false
          schema:
            type: string
            enum: ["18-30", "30-55", "55+"]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
  /scenarios/{scenarioId}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get a specific scenario by ID
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: integer
        - name: lang
          in: query
          required: false
          description: ISO 639-1 language code for scenario content localization
          schema:
            type: string
      responses:
        "200":
          description: Scenario detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Scenario:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        pillar:
          type: integer
        region_code:
          type: string
        age_group:
          type: string
        difficulty_level:
          type: integer
        estimated_minutes:
          type: integer
        noble_score_impact:
          type: integer
        is_regional:
          type: boolean
          description: True when the scenario belongs to the requested region; false when it is a universal fallback.
        content_json:
          $ref: "#/components/schemas/ScenarioContent"
      required:
        - id
        - title
        - pillar
        - region_code
        - difficulty_level
        - noble_score_impact
        - content_json
    ScenarioContent:
      type: object
      properties:
        situation:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioOption"
      required:
        - situation
        - question
        - options
    ScenarioOption:
      type: object
      properties:
        text:
          type: string
        correct:
          type: boolean
        explanation:
          type: string
      required:
        - text
        - correct
        - explanation
...